requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
pandas>=1.3.0
opencv-python-headless>=4.5.0
Pillow>=8.3.0
albumentations>=1.1.0
einops>=0.4.0
tqdm>=4.62.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0
scikit-image>=0.18.0
jupyter>=1.0.0
pyyaml>=5.4.0
